working on production

+++ Controller/Runtime/Items/Crop.cs
@@ -8,8 +8,8 @@ namespace Soul.Controller.Runtime.Items
     [CreateAssetMenu(fileName = "crop", menuName = "Soul/Item/Create Crop")]
     public class Crop : Item, IWeight
     {
-        [Range(0, 100)] [SerializeReference] public float weight = 1;
+        [Range(0, 100)] [SerializeReference] public int weight = 1;
         public UnityTimeSpan growTime;
-        public float Weight => weight;
+        public int Weight => weight;
     }
 }
